+++ sstmac/hardware/topology/butterfly.h
@@ -158,7 +158,7 @@ class butterfly :
   switch_id
   netlink_to_ejection_switch(
     node_id nodeaddr,
-    int &switch_port) const override;
+    uint16_t &switch_port) const override;
 
   void
   nodes_connected_to_injection_switch(switch_id swaddr,
