+++ sstmac/hardware/topology/topology.cc
@@ -294,12 +294,12 @@ class merlin_topology : public topology {
   }
 
   switch_id
-  netlink_to_injection_switch(node_id nodeaddr, int& switch_port) const override {
+  netlink_to_injection_switch(node_id nodeaddr, uint16_t& switch_port) const override {
     spkt_abort_printf("merlin topology functions should never be called");
   }
 
   switch_id
-  netlink_to_ejection_switch(node_id nodeaddr, int& switch_port) const override {
+  netlink_to_ejection_switch(node_id nodeaddr, uint16_t& switch_port) const override {
     spkt_abort_printf("merlin topology functions should never be called");
   }
 
@@ -309,12 +309,12 @@ class merlin_topology : public topology {
   }
 
   switch_id
-  node_to_ejection_switch(node_id addr, int& port) const override {
+  node_to_ejection_switch(node_id addr, uint16_t& port) const override {
     spkt_abort_printf("merlin topology functions should never be called");
   }
 
   switch_id
-  node_to_injection_switch(node_id addr, int& port) const override {
+  node_to_injection_switch(node_id addr, uint16_t& port) const override {
     spkt_abort_printf("merlin topology functions should never be called");
   }
